Deleted BlueDuck.jsx file and integrated in index.jsx file
