--- src/screens/BlueDuckV/BlueDuckV.jsx
@@ -1,111 +1,0 @@
-import React from "react";
-import { Navbar } from "../../components/ui/Navbar";
-import { AboutUsSection } from "./sections/Section1/HeroSection";
-import { CallToActionSection } from "./sections/Section7/CallToActionSection";
-import { ContactSection } from "./sections/Section2/About";
-import { HeroSection } from "./sections/Section4/ChooseUs";
-import { NavigationSection } from "./sections/Section3/Services";
-import { PortfolioSection } from "./sections/Section5/PortfolioSection";
-import { ServicesSection } from "./sections/Section6/ServicesSection";
-import { TestimonialsSection } from "./sections/Section8/Footer";
-
-export const BlueDuckV = () => {
-  const logoData = [
-    {
-      src: "https://c.animaapp.com/mf9h3e3wUYtRXO/img/deepseek-text-1.png",
-      alt: "Deepseek text",
-    },
-    {
-      src: "https://c.animaapp.com/mf9h3e3wUYtRXO/img/coca-cola-logo-white-1.png",
-      alt: "Coca cola logo white",
-    },
-    {
-      src: "https://c.animaapp.com/mf9h3e3wUYtRXO/img/adobe-express---file--7--1.png",
-      alt: "Adobe express file",
-    },
-    {
-      src: "https://c.animaapp.com/mf9h3e3wUYtRXO/img/avidxchange-primarylogo-rgbh-1.png",
-      alt: "Avidxchange",
-    },
-    {
-      src: "https://c.animaapp.com/mf9h3e3wUYtRXO/img/adobe-express---file--6--1.png",
-      alt: "Adobe express file",
-    },
-    {
-      src: "https://c.animaapp.com/mf9h3e3wUYtRXO/img/new-paypal-logo-white-png-horizontal-large-size-1-3.png",
-      alt: "New paypal logo",
-    },
-    {
-      src: "https://c.animaapp.com/mf9h3e3wUYtRXO/img/xpay-preto-1.png",
-      alt: "Xpay preto",
-    },
-  ];
-
-  return (
-    <div
-      className="bg-black flex flex-col items-center w-screen min-h-screen overflow-x-hidden"
-      data-model-id="3816:7750"
-    >
-      {/* ✅ Navbar on top */}
-      <Navbar />
-
-      <div className="bg-black overflow-hidden w-full max-w-[1440px] flex flex-col">
-        <section className="w-full translate-y-[-1rem] animate-fade-in opacity-0 [--animation-delay:0ms]">
-          <AboutUsSection />
-        </section>
-
-        <section className="w-full h-[758px] bg-[#c4c4c4] flex items-center justify-center translate-y-[-1rem] animate-fade-in opacity-0 [--animation-delay:200ms] mb-12">
-          <div className="[font-family:'Helvetica_Neue-Regular',Helvetica] font-normal text-black text-[17px] text-center tracking-[0] leading-[normal]">
-            Placeholder for a video that will expand with animation
-          </div>
-        </section>
-
-        {/* ✅ Logos Section (responsive & images fit screen) */}
-        <section className="w-full flex justify-center py-8 translate-y-[-1rem] animate-fade-in opacity-0 [--animation-delay:400ms] mb-24">
-          <div className="flex flex-wrap items-center justify-center gap-12 opacity-80 px-6">
-            {logoData.map((logo, index) => (
-              <div
-                key={`logo-${index}`}
-                className="flex items-center justify-center max-h-[46px]"
-              >
-                <img
-                  className="max-h-[46px] w-auto object-contain"
-                  alt={logo.alt}
-                  src={logo.src}
-                />
-              </div>
-            ))}
-          </div>
-        </section>
-
-        <section className="w-full translate-y-[-1rem] animate-fade-in opacity-0 [--animation-delay:600ms]">
-          <ContactSection />
-        </section>
-
-        <section className="w-full translate-y-[-1rem] animate-fade-in opacity-0 [--animation-delay:800ms]">
-          <NavigationSection />
-        </section>
-
-        <section className="w-full translate-y-[-1rem] animate-fade-in opacity-0 [--animation-delay:1000ms]">
-          <HeroSection />
-        </section>
-
-        <section className="w-full translate-y-[-1rem] animate-fade-in opacity-0 [--animation-delay:1200ms]">
-          <PortfolioSection />
-        </section>
-
-        <section className="w-full translate-y-[-1rem] animate-fade-in opacity-0 [--animation-delay:1400ms]">
-          <ServicesSection />
-        </section>
-
-        <section className="w-full translate-y-[-1rem] animate-fade-in opacity-0 [--animation-delay:1600ms]">
-          <CallToActionSection />
-        </section>
-
-        <section className="w-full translate-y-[-1rem] animate-fade-in opacity-0 [--animation-delay:1800ms]">
-          <TestimonialsSection />
-        </section>
-      </div>
-    </div>
-  );
-};
